[IMP] - tehtavat 1,2,3 (ja uusi yritys 1 ja 2 osalta)
