+++ kannatusseura/__manifest__.py
@@ -11,8 +11,10 @@
     'depends': ['contacts'],
     'data': [
         'security/ir.model.access.csv',
-        'views/res_partner_views.xml',
+        'views/kannatusseura_views.xml',
         'views/kannatusseura_cat.xml',
+        'views/res_partner_views.xml',
+
     ],
     'demo': [
         'demo/demo.xml',
