+++ kannatusseura/security/ir.model.access.csv
@@ -1,2 +1,3 @@
 id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
 access_kannatusseura_cat,kannatusseura.cat,model_kannatusseura_cat,,1,1,1,1
+access_kannatusseura,kannatusseura.model,model_kannatusseura_model,,1,1,1,1
