+++ kannatusseura/views/kannatusseura_cat.xml
@@ -1,28 +1,27 @@
 <odoo>
-    <record id="view_kannatusseura_cat_form" model="ir.ui.view">
-        <field name="name">kannatusseura.cat.form</field>
-        <field name="model">kannatusseura.cat</field>
-        <field name="arch" type="xml">
-            <form>
-                <group>
-                    <field name="kannatusseura_cat_name"  />
-                </group>
+    <data>
+        <record id="view_kannatusseura_cat_form" model="ir.ui.view">
+            <field name="name">kannatusseura.cat.form</field>
+            <field name="model">kannatusseura.cat</field>
+            <field name="arch" type="xml">
+                <form>
+                    <group>
+                        <field name="kannatusseura_cat_name"  string="Kannatusseura category name" />
+                        <field name="kannatusseura_cat_code"  string="Kannatusseura category code" />
+                    </group>
 
-            </form>
-        </field>
-    </record>
+                </form>
+            </field>
+        </record>
 
-    <record id="kcat_hockey" model="kannatusseura.cat">
-        <field name="kannatusseura_cat_name">hockey</field>
-        <field name="kannatusseura_cat_code">HCK</field>
-    </record>
-    <record id="kcat_soccer" model="kannatusseura.cat">
-        <field name="kannatusseura_cat_name">soccer</field>
-        <field name="kannatusseura_cat_code">SOC</field>
-    </record>
-    <record id="kcat_formula1" model="kannatusseura.cat">
-        <field name="kannatusseura_cat_name">formula1</field>
-        <field name="kannatusseura_cat_code">FRM</field>
-    </record>
+
+        <record id="action_kannatusseura_cat" model="ir.actions.act_window">
+            <field name="name">Kannatusseura Categories</field>
+            <field name="res_model">kannatusseura.cat</field>
+            <field name="view_mode">tree,form</field>
+        </record>
+        <!-- <menuitem id="menu_kannatusseura_cat" name="Categories" parent="menu_kannatusseura_root" action="action_kannatusseura_cat"/> -->
+
+    </data>
 
 </odoo>
